refactor moved repositories to `/repository/gorm_postgres/`
