+++ internal/repository/gorm_postgres/db.go
@@ -1,4 +1,4 @@
-package repository
+package gorm_postgres
 
 import (
 	"gorm.io/driver/postgres"
