+++ internal/repository/gorm_postgres/mark_repo.go
@@ -1,4 +1,4 @@
-package repository
+package gorm_postgres
 
 import (
 	"context"
