+++ internal/repository/gorm_postgres/note_repo.go
@@ -1,4 +1,4 @@
-package repository
+package gorm_postgres
 
 import (
 	"context"
